+++ src/vr180_convert/remapper/radius.py
@@ -4,8 +4,8 @@ import attrs
 import ivy
 from ivy import Array
 
+from .base import RemapperBase
 from .denormalize import DenormalizeRemapper
-from .remapper import RemapperBase
 
 
 def _get_radius(input: Array, *, threshold: int = 10) -> float:
